requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
google-generativeai==0.8.3
gspread==6.1.4
google-auth==2.35.0
requests==2.32.3
python-dateutil==2.9.0.post0
